Chore: Reservation 필드 수정
